# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "商城前端与运营中台共用的后端接口。响应统一信封 {code,message,data}；\n认证为 HttpOnly Cookie sso_token，写操作需回传 X-CSRF-Token。\n文档生成：swag init -g main.go -o docs/swagger，访问 /swagger/index.html",
        "title": "FreshMart 生鲜商城 API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/api/auth/password/reset": {
            "post": {
                "description": "邮箱不存在同样返回成功，避免账号枚举",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "发起找回密码",
                "parameters": [
                    {
                        "description": "邮箱",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/password/reset-verify": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "校验重置验证码",
                "parameters": [
                    {
                        "description": "验证码与新密码",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ResetVerifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/api/login": {
            "post": {
                "description": "账号可以是 SSO 用户名，也可以是商城会员邮箱/用户名/手机号；成功后写入 sso_token 与 sso_csrf",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "登录并签发会话 Cookie",
                "parameters": [
                    {
                        "description": "账号与密码",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "用户名或密码错误",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/api/miniprogram/bind": {
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "已登录会员（密码登录或 H5 会话）用 wx.login 的 code 把 openid 绑到自己账号上，绑完即可静默登录\n只接受 code：openid 由服务端重新换取，不接受客户端自报，否则可以把别人的收款身份挂到自己账号上",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "miniprogram"
                ],
                "summary": "绑定微信身份",
                "parameters": [
                    {
                        "description": "wx.login 的 code",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MpBindRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "缺少 code 或 code 无效",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "401": {
                        "description": "未登录",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "409": {
                        "description": "该微信已绑定其他账号",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    }
                }
            }
        },
        "/api/miniprogram/wechat/login": {
            "post": {
                "description": "用 wx.login 的 code 换 openid；openid 未入库时按平台授权的手机号合并或新建账号，未带 phone_code 时不发令牌（422）\n响应不回 Cookie，客户端自行持有 token；请求 URL 上带 X-Client: mp_wechat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "miniprogram"
                ],
                "summary": "微信小程序授权登录",
                "parameters": [
                    {
                        "description": "平台凭证 code 与可选的手机号授权 code",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.MpWechatLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "缺少 code 或 code 无效",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "409": {
                        "description": "该手机号已绑定其他微信",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "422": {
                        "description": "未授权手机号，客户端应拉起授权后重试",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    }
                }
            }
        },
...
        "/api/register": {
            "post": {
                "description": "校验口径与前端 validateProfileField 一致；密码以 bcrypt 存储；account_type 为 merchant 时同步建门店并把类型落库到账号",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "account"
                ],
                "summary": "注册商城会员",
                "parameters": [
                    {
                        "description": "注册信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "400": {
                        "description": "字段校验失败",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    },
                    "409": {
                        "description": "邮箱或手机号已被注册",
                        "schema": {
                            "$ref": "#/definitions/models.ApiResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ApiResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "HTTP 状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "业务数据，类型根据接口而定",
                    "type": "object"
                },
                "message": {
                    "description": "提示信息，如 \"success\" 或错误描述",
                    "type": "string"
                }
            }
        },
...
        "models.LoginRequest": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "email": {
                    "description": "邮箱，语义同上",
                    "type": "string"
                },
                "password": {
                    "description": "密码，必填",
                    "type": "string"
                },
                "username": {
                    "description": "用户名、邮箱或手机号",
                    "type": "string"
                }
            }
        },
        "models.MpBindRequest": {
            "type": "object",
            "required": [
                "code"
            ],
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        },
        "models.MpWechatLoginRequest": {
            "type": "object",
            "required": [
                "code"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "phone_code": {
                    "type": "string"
                }
            }
        },
...
        "models.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "mobile",
                "password",
                "username"
            ],
            "properties": {
                "account_type": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "mobile": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "store_address": {
                    "type": "string"
                },
                "store_name": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.ResetPasswordRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "models.ResetVerifyRequest": {
            "type": "object",
            "required": [
                "code",
                "new_password"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "sso_token",
            "in": "cookie"
        }
    }
}